+++ KittsMenuSystem/Features/Settings/ABButton.cs
@@ -18,14 +18,14 @@ namespace KittsMenuSystem.Features.Settings;
 /// <param name="defaultIsB">Sets <see cref="SSTwoButtonsSetting.DefaultIsB"/>.</param>
 /// <param name="hint">Hint of <see cref="SSTwoButtonsSetting"/>.</param>
 public class ABButton(int? id, string label, string optionA, string optionB, Action<ReferenceHub, bool, SSTwoButtonsSetting> onChanged = null, bool defaultIsB = false, string hint = null)
-    : BaseSetting(new SSTwoButtonsSetting(SetValidId(id, label), label, optionA, optionB, defaultIsB, hint))
+    : BaseSetting(new SSTwoButtonsSetting(id ?? Guid.NewGuid().ToString().GetStableHashCode(), label, optionA, optionB, defaultIsB, hint))
 {
     /// <summary>
     /// Initialize new <see cref="ABButton"/> setting (automatic id) with base <see cref="SSTwoButtonsSetting"/> that calls <see cref="Action"/> when changed.
     /// </summary>
     /// <param name="label">Label of <see cref="SSTwoButtonsSetting"/>.</param>
-    /// <param name="optionA">Label of <see cref="SSTwoButtonsSetting.OptionA"/>.</param>
-    /// <param name="optionB">Label of <see cref="SSTwoButtonsSetting.OptionB"/>.</param>
+    /// <param name="optionA">Text of <see cref="SSTwoButtonsSetting.OptionA"/>.</param>
+    /// <param name="optionB">Text of <see cref="SSTwoButtonsSetting.OptionB"/>.</param>
     /// <param name="onChanged">Tiggers <see cref="Action"/> when <see cref="ABButton"/> changes.</param>
     /// <param name="defaultIsB">Sets <see cref="SSTwoButtonsSetting.DefaultIsB"/>.</param>
     /// <param name="hint">Hint of <see cref="SSTwoButtonsSetting"/>.</param>
@@ -41,8 +41,34 @@ public class ABButton(int? id, string label, string optionA, string optionB, Act
     public Action<ReferenceHub, bool, SSTwoButtonsSetting> OnChanged { get; } = onChanged;
 
     /// <summary>
-    /// Makes sure the Id is valid (not null)
+    /// Shortcut to underlying <see cref="SSTwoButtonsSetting.OptionA"/>.
     /// </summary>
-    internal static int SetValidId(int? id, string label) =>
-        id ?? (label + "ABButton").GetStableHashCode();
+    public string OptionA
+    {
+        get => (Base as SSTwoButtonsSetting).OptionA;
+        set => UpdateOptionA(value);
+    }
+
+    /// <summary>
+    /// Shortcut to underlying <see cref="SSTwoButtonsSetting.SendTwoButtonUpdate"/> but only for <see cref="SSTwoButtonsSetting.OptionA"/>.
+    /// <see cref="SSTwoButtonsSetting.OptionA"/> will go back to default (what was set in the menu) when rejoining server.
+    /// </summary>
+    public void UpdateOptionA(string newOptionA, bool applyOverride = true, Func<ReferenceHub, bool> receiveFilter = null) =>
+        (Base as SSTwoButtonsSetting).SendTwoButtonUpdate(newOptionA, OptionB, applyOverride, receiveFilter);
+
+    /// <summary>
+    /// Shortcut to underlying <see cref="SSTwoButtonsSetting.OptionB"/>.
+    /// </summary>
+    public string OptionB
+    {
+        get => (Base as SSTwoButtonsSetting).OptionB;
+        set => UpdateOptionB(value);
+    }
+
+    /// <summary>
+    /// Shortcut to underlying <see cref="SSTwoButtonsSetting.SendTwoButtonUpdate"/> but only for <see cref="SSTwoButtonsSetting.OptionB"/>.
+    /// <see cref="SSTwoButtonsSetting.OptionB"/> will go back to default (what was set in the menu) when rejoining server.
+    /// </summary>
+    public void UpdateOptionB(string newOptionB, bool applyOverride = true, Func<ReferenceHub, bool> receiveFilter = null) =>
+        (Base as SSTwoButtonsSetting).SendTwoButtonUpdate(OptionA, newOptionB, applyOverride, receiveFilter);
 }