+++ KittsMenuSystem/Features/Settings/Button.cs
@@ -12,12 +12,12 @@ namespace KittsMenuSystem.Features.Settings;
 /// </remarks>
 /// <param name="id">Id of <see cref="SSButton"/>.</param>
 /// <param name="label">Label of <see cref="SSButton"/>.</param>
-/// <param name="buttonText">Text of <see cref="SSButton"/>.</param>
+/// <param name="buttonText">Text of <see cref="SSButton.ButtonText"/>.</param>
 /// <param name="onPressed">Triggers <see cref="Action"/> when <see cref="Button"/> pressed.</param>
 /// <param name="holdTimeSeconds">Sets <see cref="SSButton.HoldTimeSeconds"/>.</param>
 /// <param name="hint">Hint of <see cref="SSButton"/>.</param>
 public class Button(int? id, string label, string buttonText, Action<ReferenceHub, SSButton> onPressed = null, float? holdTimeSeconds = null, string hint = null)
-    : BaseSetting(new SSButton(SetValidId(id, label), label, buttonText, holdTimeSeconds, hint))
+    : BaseSetting(new SSButton(id ?? Guid.NewGuid().ToString().GetStableHashCode(), label, buttonText, holdTimeSeconds, hint))
 {
     /// <summary>
     /// Initialize new <see cref="Button"/> setting (automatic id) with base <see cref="SSButton"/> that calls <see cref="Action"/> when pressed.
@@ -38,8 +38,34 @@ public class Button(int? id, string label, string buttonText, Action<ReferenceHu
     public Action<ReferenceHub, SSButton> OnPressed { get; } = onPressed;
 
     /// <summary>
-    /// Makes sure the Id is valid (not null)
+    /// Shortcut to underlying <see cref="SSButton.ButtonText"/>.
     /// </summary>
-    internal static int SetValidId(int? id, string label) =>
-        id ?? (label + "Button").GetStableHashCode();
+    public string ButtonText
+    {
+        get => (Base as SSButton).ButtonText;
+        set => UpdateButtonText(value);
+    }
+
+    /// <summary>
+    /// Shortcut to underlying <see cref="SSButton.SendButtonUpdate"/> but only for <see cref="SSButton.ButtonText"/>.
+    /// <see cref="SSButton.ButtonText"/> will go back to default (what was set in the menu) when rejoining server.
+    /// </summary>
+    public void UpdateButtonText(string newButtonText, bool applyOverride = true, Func<ReferenceHub, bool> receiveFilter = null) =>
+        (Base as SSButton).SendButtonUpdate(newButtonText, HoldTimeSeconds, applyOverride, receiveFilter);
+
+    /// <summary>
+    /// Shortcut to underlying <see cref="SSButton.HoldTimeSeconds"/>.
+    /// </summary>
+    public float HoldTimeSeconds
+    {
+        get => (Base as SSButton).HoldTimeSeconds;
+        set => UpdateHoldTimeSeconds(value);
+    }
+
+    /// <summary>
+    /// Shortcut to underlying <see cref="SSButton.SendButtonUpdate"/> but only for <see cref="SSButton.HoldTimeSeconds"/>.
+    /// <see cref="SSButton.HoldTimeSeconds"/> will go back to default (what was set in the menu) when rejoining server.
+    /// </summary>
+    public void UpdateHoldTimeSeconds(float newHoldTimeSeconds, bool applyOverride = true, Func<ReferenceHub, bool> receiveFilter = null) =>
+        (Base as SSButton).SendButtonUpdate(ButtonText, newHoldTimeSeconds, applyOverride, receiveFilter);
 }